chore: Break into functions
